Fix site structure
