--- ExampleSite/CommunityKitSite/.gitignore
@@ -1,9 +1,0 @@
-.DS_Store
-Build
-.build
-Packages
-xcuserdata/
-DerivedData/
-.swiftpm/configuration/registries.json
-.swiftpm/xcode/package.xcworkspace/contents.xcworkspacedata
-.netrc
