--- ExampleSite/CommunityKitSite/Assets/Assets.txt
@@ -1,1 +1,0 @@
-Place any assets you want copied into the build here.
