--- ExampleSite/CommunityKitSite/Assets/images/Images.txt
@@ -1,1 +1,0 @@
-This is the best place to put images for your build, creating subfolders below here as needed.
