--- ExampleSite/CommunityKitSite/Package.swift
@@ -1,17 +1,0 @@
-// swift-tools-version: 5.9
-// The swift-tools-version declares the minimum version of Swift required to build this package.
-
-import PackageDescription
-
-let package = Package(
-    name: "IgniteStarter",
-    platforms: [.macOS(.v13)],
-    dependencies: [
-        .package(url: "https://github.com/twostraws/Ignite.git", branch: "main")
-    ],
-    targets: [
-        .executableTarget(
-            name: "IgniteStarter",
-            dependencies: ["Ignite"]),
-    ]
-)
