--- ExampleSite/CommunityKitSite/README.md
@@ -1,5 +1,0 @@
-# IgniteStarter
-
-This repository contains a starter project built using the [Ignite static site generator](https://github.com/twostraws/Ignite) – it includes just enough to get you going, so you should refer to the [IgniteSamples](https://github.com/twostraws/IgniteSamples) repository for detailed code samples.
-
-You'll find small text file in each folder explaining what that folder does. Once you're happy with that, you should delete those files.
