--- ExampleSite/CommunityKitSite/Sources/Layouts/MainLayout.swift
@@ -1,11 +1,0 @@
-import Foundation
-import Ignite
-
-struct MainLayout: Layout {
-    var body: some HTML {
-        Body {
-            content
-            IgniteFooter()
-        }
-    }
-}
